pull roster related things out of teacher sections redux
