+++ apps/src/templates/teacherDashboard/SectionActionDropdown.jsx
@@ -10,8 +10,7 @@ import {
   sectionCode,
   sectionName,
   removeSection,
-  toggleSectionHidden,
-  importOrUpdateRoster
+  toggleSectionHidden
 } from './teacherSectionsRedux';
 import {connect} from 'react-redux';
 import PrintCertificates from './PrintCertificates';
@@ -21,7 +20,10 @@ import Button from '../Button';
 import DialogFooter from './DialogFooter';
 import QuickActionsCell from '@cdo/apps/templates/tables/QuickActionsCell';
 import {getStore} from '@cdo/apps/redux';
-import {setRosterProvider} from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
+import {
+  setRosterProvider,
+  importOrUpdateRoster
+} from '@cdo/apps/code-studio/rosterRedux';
 import {SectionLoginType} from '@cdo/apps/util/sharedConstants';
 
 class SectionActionDropdown extends Component {
