+++ apps/test/unit/templates/teacherDashboard/SectionActionDropdownTest.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {shallow} from 'enzyme';
 import {expect} from '../../../util/deprecatedChai';
 import {UnconnectedSectionActionDropdown as SectionActionDropdown} from '@cdo/apps/templates/teacherDashboard/SectionActionDropdown';
-import {setRosterProvider} from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
+import {setRosterProvider} from '@cdo/apps/code-studio/rosterRedux';
 import PrintCertificates from '@cdo/apps/templates/teacherDashboard/PrintCertificates';
 
 const sections = [
